Conflicts:
	tools/benchmark/benchmark.go

+++ batch_command.go
@@ -90,7 +90,7 @@ func (cmd *baseMultiCommand) parseKey(fieldCount int) (*Key, error) {
 			return nil, err
 		}
 
-		fieldlen := int(Buffer.BytesToInt32(cmd.dataBuffer, 0))
+		fieldlen := int(uint32(Buffer.BytesToInt32(cmd.dataBuffer, 0)))
 		if err = cmd.readBytes(fieldlen); err != nil {
 			return nil, err
 		}
