+++ query_record_command.go
@@ -60,10 +60,10 @@ func (cmd *queryRecordCommand) parseRecordResults(ifc command, receiveSize int) 
 			return false, nil
 		}
 
-		generation := int(Buffer.BytesToInt32(cmd.dataBuffer, 6))
-		expiration := int(Buffer.BytesToInt32(cmd.dataBuffer, 10))
-		fieldCount := int(Buffer.BytesToInt16(cmd.dataBuffer, 18))
-		opCount := int(Buffer.BytesToInt16(cmd.dataBuffer, 20))
+		generation := int(uint32(Buffer.BytesToInt32(cmd.dataBuffer, 6)))
+		expiration := int(uint32(Buffer.BytesToInt32(cmd.dataBuffer, 10)))
+		fieldCount := int(uint16(Buffer.BytesToInt16(cmd.dataBuffer, 18)))
+		opCount := int(uint16(Buffer.BytesToInt16(cmd.dataBuffer, 20)))
 
 		key, err := cmd.parseKey(fieldCount)
 		if err != nil {
@@ -80,7 +80,7 @@ func (cmd *queryRecordCommand) parseRecordResults(ifc command, receiveSize int) 
 				return false, err
 			}
 
-			opSize := int(Buffer.BytesToInt32(cmd.dataBuffer, 0))
+			opSize := int(uint32(Buffer.BytesToInt32(cmd.dataBuffer, 0)))
 			particleType := int(cmd.dataBuffer[5])
 			nameSize := int(cmd.dataBuffer[7])
